+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### unreleased
+* Remove dependency on unittest2, use unittest instead (#113)
+
 ### 3.4.2 - 2018-08-21
 
 * Fix typos in our warning messages https://github.com/quandl/quandl-python/pull/114